config final docker

+++ production-backend/src/main/java/com/production/backend/repository/ResourceRepository.java
@@ -3,5 +3,5 @@ package com.production.backend.repository;
 import com.production.backend.entity.Resource;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ResourceRepository extends JpaRepository<Resource, Integer> {
+public interface ResourceRepository extends JpaRepository<Resource, Long> {
 }
